requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
email-validator>=2.1.0.post1
